+++ src/modules/trial-and-error/Mb.cpp
@@ -917,6 +917,7 @@ struct MbWidget : ModuleWidget {
 			[]() { pluginSettings.mbMagnifierEnabled ^= true; }
 		));
 		menu->addChild(createBoolPtrMenuItem("Apply VCV Libray Whitelist", "", &pluginSettings.mbApplyLibraryWhitelist));
+		menu->addChild(createBoolPtrMenuItem("Show deprecated models", "", &pluginSettings.mbShowDeprecated));
 
 		menu->addChild(new MenuSeparator());
 		menu->addChild(createMenuLabel("Custom tags"));