+++ src/pluginsettings.hpp
@@ -15,6 +15,7 @@ struct Settings {
 	float mbSearchThreshold = 0.5f;
 	bool mbMagnifierEnabled = false;
 	bool mbApplyLibraryWhitelist = false;
+	bool mbShowDeprecated = false;
 
 	NVGcolor overlayTextColor = bndGetTheme()->menuTheme.textColor;
 	int overlayHpos = 0;