+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Changes and Fixes
 
 - Module [MB](./docs/trial-and-error/Mb.md)
-    - Added option to apply VCV Library Whitelisting
+    - Added option to apply VCV Library whitelisting
+    - Added option to show deprecated module models (now hidden by default) (#440)
 
 ## 2.4.0